2020.03.09 작업 : 블로그 2종 추가, 파일이 존재할 경우 이어쓰기 처리

+++ src/eia/beffatStrtgySmallScaleDscssSttusInfoInqireService/GetBsnsStrtgySmallScaleDscssBsnsDetailInfoInqire.java
@@ -35,37 +35,45 @@ public class GetBsnsStrtgySmallScaleDscssBsnsDetailInfoInqire {
 			// step 1.파일의 첫 행 작성
 			File file = new File(JsonParser.getProperty("file_path") + "EIA/TIF_EIA_54.dat");
 
-			try {
+			if(file.exists()){
+				
+				System.out.println("파일이 이미 존재하므로 이어쓰기..");
+				
+			} else {
+			
+				try {
 
-				PrintWriter pw = new PrintWriter(new BufferedWriter(new FileWriter(file, true)));
-				pw.write("resultCode"); // 결과코드
-				pw.write("|^");
-				pw.write("resultMsg"); // 결과메시지
-				pw.write("|^");
-				pw.write("perCd"); // 사전환경성검토코드
-				pw.write("|^");
-				pw.write("bizNm"); // 사업명
-				pw.write("|^");
-				pw.write("ccilOrganCd"); // 협의기관
-				pw.write("|^");
-				pw.write("ctcMemNm"); // 담당자
-				pw.write("|^");
-				pw.write("ctcMemTeam"); // 담당부서
-				pw.write("|^");
-				pw.write("ctcMemEmail"); // E-mail
-				pw.write("|^");
-				pw.write("ctcMemTel"); // 전화번호
-				pw.write("|^");
-				pw.write("ctcMemFax"); // Fax 번호
-				pw.write("|^");
-				pw.write("ccilResFl"); // 협의결과
+					PrintWriter pw = new PrintWriter(new BufferedWriter(new FileWriter(file, true)));
+					pw.write("resultCode"); // 결과코드
+					pw.write("|^");
+					pw.write("resultMsg"); // 결과메시지
+					pw.write("|^");
+					pw.write("perCd"); // 사전환경성검토코드
+					pw.write("|^");
+					pw.write("bizNm"); // 사업명
+					pw.write("|^");
+					pw.write("ccilOrganCd"); // 협의기관
+					pw.write("|^");
+					pw.write("ctcMemNm"); // 담당자
+					pw.write("|^");
+					pw.write("ctcMemTeam"); // 담당부서
+					pw.write("|^");
+					pw.write("ctcMemEmail"); // E-mail
+					pw.write("|^");
+					pw.write("ctcMemTel"); // 전화번호
+					pw.write("|^");
+					pw.write("ctcMemFax"); // Fax 번호
+					pw.write("|^");
+					pw.write("ccilResFl"); // 협의결과
 
-				pw.println();
-				pw.flush();
-				pw.close();
+					pw.println();
+					pw.flush();
+					pw.close();
 
-			} catch (IOException e) {
-				e.printStackTrace();
+				} catch (IOException e) {
+					e.printStackTrace();
+				}
+			
 			}
 
 			String json = "";
