+++ src/eia/envrnAffcEvlDecsnCnInfoInqireService/GetComCodeInfoInqire.java
@@ -34,29 +34,38 @@ public class GetComCodeInfoInqire {
 			// step 1.파일의 첫 행 작성
 			File file = new File(JsonParser.getProperty("file_path") + "EIA/TIF_EIA_48.dat");
 
-			try {
+			if(file.exists()){
+				
+				System.out.println("파일이 이미 존재하므로 이어쓰기..");
+				
+			} else {
+			
+				try {
 
-				PrintWriter pw = new PrintWriter(new BufferedWriter(new FileWriter(file, true)));
-				pw.write("resultCode"); // 결과코드
-				pw.write("|^");
-				pw.write("resultMsg"); // 결과메시지
-				pw.write("|^");
-				pw.write("comCd"); // 공통코드
-				pw.write("|^");
-				pw.write("jongCd"); // 코드종류
-				pw.write("|^");
-				pw.write("cdNm"); // 코드명
-				pw.write("|^");
-				pw.write("gubunFl"); // 구분상태
-				pw.write("|^");
-				pw.write("sortSeq"); // 소팅 일련번호
-				pw.println();
-				pw.flush();
-				pw.close();
+					PrintWriter pw = new PrintWriter(new BufferedWriter(new FileWriter(file, true)));
+					pw.write("resultCode"); // 결과코드
+					pw.write("|^");
+					pw.write("resultMsg"); // 결과메시지
+					pw.write("|^");
+					pw.write("comCd"); // 공통코드
+					pw.write("|^");
+					pw.write("jongCd"); // 코드종류
+					pw.write("|^");
+					pw.write("cdNm"); // 코드명
+					pw.write("|^");
+					pw.write("gubunFl"); // 구분상태
+					pw.write("|^");
+					pw.write("sortSeq"); // 소팅 일련번호
+					pw.println();
+					pw.flush();
+					pw.close();
 
-			} catch (IOException e) {
-				e.printStackTrace();
+				} catch (IOException e) {
+					e.printStackTrace();
+				}
+			
 			}
+			
 
 			String json = "";
