+++ src/eia/envrnAffcEvlDscssSttusInfoInqireService/GetDscssBsnsListInfoInqire.java
@@ -35,38 +35,46 @@ public class GetDscssBsnsListInfoInqire {
 			// step 1.파일의 첫 행 작성
 			File file = new File(JsonParser.getProperty("file_path") + "EIA/TIF_EIA_41.dat");
 
-			try {
+			if(file.exists()){
+				
+				System.out.println("파일이 이미 존재하므로 이어쓰기..");
+				
+			} else {
+			
+				try {
 
-				PrintWriter pw = new PrintWriter(new BufferedWriter(new FileWriter(file, true)));
-				pw.write("resultCode"); // 결과코드
-				pw.write("|^");
-				pw.write("resultMsg"); // 결과메세지
-				pw.write("|^");
-				pw.write("numOfRows"); // 한 페이지 결과 수
-				pw.write("|^");
-				pw.write("pageNo"); // 페이지 번호
-				pw.write("|^");
-				pw.write("totalCount"); // 전체 결과 수
-				pw.write("|^");
-				pw.write("rnum"); // 정렬순서
-				pw.write("|^");
-				pw.write("eiaCd"); // 환경영향평가 코드
-				pw.write("|^");
-				pw.write("eiaSeq"); // 환경영향평가 고유번호
-				pw.write("|^");
-				pw.write("bizNm"); // 사업명
-				pw.write("|^");
-				pw.write("ccilOrganNm"); // 협의기관
-				pw.write("|^");
-				pw.write("firstCtgCd"); // 현재단계
-				pw.write("|^");
-				pw.write("stepChangeDt"); // 단계변경일
-				pw.println();
-				pw.flush();
-				pw.close();
+					PrintWriter pw = new PrintWriter(new BufferedWriter(new FileWriter(file, true)));
+					pw.write("resultCode"); // 결과코드
+					pw.write("|^");
+					pw.write("resultMsg"); // 결과메세지
+					pw.write("|^");
+					pw.write("numOfRows"); // 한 페이지 결과 수
+					pw.write("|^");
+					pw.write("pageNo"); // 페이지 번호
+					pw.write("|^");
+					pw.write("totalCount"); // 전체 결과 수
+					pw.write("|^");
+					pw.write("rnum"); // 정렬순서
+					pw.write("|^");
+					pw.write("eiaCd"); // 환경영향평가 코드
+					pw.write("|^");
+					pw.write("eiaSeq"); // 환경영향평가 고유번호
+					pw.write("|^");
+					pw.write("bizNm"); // 사업명
+					pw.write("|^");
+					pw.write("ccilOrganNm"); // 협의기관
+					pw.write("|^");
+					pw.write("firstCtgCd"); // 현재단계
+					pw.write("|^");
+					pw.write("stepChangeDt"); // 단계변경일
+					pw.println();
+					pw.flush();
+					pw.close();
 
-			} catch (IOException e) {
-				e.printStackTrace();
+				} catch (IOException e) {
+					e.printStackTrace();
+				}
+			
 			}
 
 			// step 2. 전체 데이터 숫자 파악을 위해 페이지 수 0으로 파싱
